refactor: remove v1 operate types
