+++ operate/client/src/modules/queries/elementInstancesStatistics/useElementStatistics.ts
@@ -11,7 +11,7 @@ import {type GetProcessInstanceStatisticsResponseBody} from '@camunda/camunda-ap
 import {getStatisticsByElement} from 'modules/utils/statistics/elementInstances';
 import {useBusinessObjects} from '../processDefinitions/useBusinessObjects';
 import type {BusinessObjects} from 'bpmn-js/lib/NavigatedViewer';
-import type {ElementState} from 'modules/types/operate';
+import type {ElementState} from 'modules/bpmn-js/overlayTypes';
 
 const statisticsByElementParser =
   (businessObjects?: BusinessObjects) =>
