+++ operate/client/src/modules/tracking/index.ts
@@ -10,13 +10,10 @@ import type {Mixpanel} from 'mixpanel-browser';
 import {getStage} from './getStage';
 import {getClientConfig} from 'modules/utils/getClientConfig';
 import type {
-  InstanceEntityState,
-  OperationEntityType,
-} from 'modules/types/operate';
-import type {
   BatchOperationState,
   BatchOperationType,
   DecisionInstanceState,
+  ProcessInstanceState,
 } from '@camunda/camunda-api-zod-schemas/8.10';
 import {useCurrentPage} from 'modules/hooks/useCurrentPage';
 
@@ -61,11 +58,11 @@ type Events =
     }
   | {
       eventName: 'batch-operation';
-      operationType: OperationEntityType;
+      operationType: BatchOperationType;
     }
   | {
       eventName: 'single-operation';
-      operationType: OperationEntityType;
+      operationType: BatchOperationType;
       source: 'instances-list' | 'incident-table' | 'instance-header';
     }
   | {
@@ -93,7 +90,7 @@ type Events =
     }
   | {
       eventName: 'process-instance-details-loaded';
-      state: InstanceEntityState;
+      state: ProcessInstanceState;
     }
   | {
       eventName: 'decision-instance-details-loaded';
