--- operate/client/src/modules/types/operate.ts
@@ -1,47 +1,0 @@
-/*
- * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH under
- * one or more contributor license agreements. See the NOTICE file distributed
- * with this work for additional information regarding copyright ownership.
- * Licensed under the Camunda License 1.0. You may not use this file
- * except in compliance with the Camunda License 1.0.
- */
-
-type OperationEntityType =
-  | 'RESOLVE_INCIDENT'
-  | 'CANCEL_PROCESS_INSTANCE'
-  | 'DELETE_PROCESS_INSTANCE'
-  | 'UPDATE_VARIABLE'
-  | 'ADD_VARIABLE'
-  | 'MODIFY_PROCESS_INSTANCE'
-  | 'DELETE_DECISION_DEFINITION'
-  | 'DELETE_PROCESS_DEFINITION'
-  | 'MIGRATE_PROCESS_INSTANCE'
-  | 'MOVE_TOKEN';
-
-type ElementState = 'active' | 'incidents' | 'canceled' | 'completed';
-
-type InstanceEntityState =
-  | 'ACTIVE'
-  | 'COMPLETED'
-  | 'CANCELED'
-  | 'INCIDENT'
-  | 'TERMINATED';
-
-type ElementInstancePlaceholder = {
-  id: string;
-  type: string;
-  state?: InstanceEntityState;
-  flowNodeId: string;
-  startDate: string;
-  endDate: null | string;
-  treePath: string;
-  sortValues: [string, string] | [];
-  isPlaceholder?: boolean;
-};
-
-export type {
-  OperationEntityType,
-  ElementState,
-  InstanceEntityState,
-  ElementInstancePlaceholder,
-};
